+++ lib/lightning/workflows/trigger.ex
@@ -42,9 +42,7 @@ defmodule Lightning.Workflows.Trigger do
     field :cron_expression, :string
     field :enabled, :boolean, default: false
 
-    field :webhook_reply, Ecto.Enum,
-      values: @webhook_reply_types,
-      default: :before_start
+    field :webhook_reply, Ecto.Enum, values: @webhook_reply_types
 
     belongs_to :workflow, Workflow
 
@@ -126,17 +124,20 @@ defmodule Lightning.Workflows.Trigger do
         changeset
         |> put_change(:cron_expression, nil)
         |> put_change(:kafka_configuration, nil)
+        |> put_default(:webhook_reply, :before_start)
 
       :cron ->
         changeset
         |> put_default(:cron_expression, "0 0 * * *")
         |> validate_cron()
         |> put_change(:kafka_configuration, nil)
+        |> put_change(:webhook_reply, nil)
 
       :kafka ->
         changeset
         |> put_change(:cron_expression, nil)
         |> validate_required([:kafka_configuration])
+        |> put_change(:webhook_reply, nil)
 
       nil ->
         changeset